apparently stuff changed

+++ utils/metrics.py
@@ -70,8 +70,8 @@ def softmax_crossentropy_loss(y_true,y_pred):
 
 
 def regression_mean_squared_loss(y_true,y_pred):
-    y_pred = tf.to_int32(K.reshape(y_pred, (-1, K.int_shape(y_pred)[-1])))
-    y_true = tf.to_int32(K.reshape(y_true, (-1, K.int_shape(y_pred)[-1])))
+    y_pred = tf.to_float(K.reshape(y_pred, (-1, K.int_shape(y_pred)[-1])))
+    y_true = tf.to_float(K.reshape(y_true, (-1, K.int_shape(y_pred)[-1])))
     sqdiff = K.square(y_pred - y_true)
     mse = K.mean(sqdiff)
     return mse
